create common test utils
