+++ app_dart/test/foundation/utils_test.dart
@@ -16,6 +16,7 @@ import '../src/datastore/fake_cocoon_config.dart';
 import '../src/request_handling/fake_http.dart';
 import '../src/request_handling/fake_logging.dart';
 import '../src/service/fake_github_service.dart';
+import '../src/utils.dart';
 
 const String branchRegExp = '''
       master
@@ -125,7 +126,3 @@ void main() {
     });
   });
 }
-
-class MockGitHub extends Mock implements GitHub {}
-
-class MockRepositoriesService extends Mock implements RepositoriesService {}
