+++ app_dart/test/request_handlers/refresh_chromebot_status_test.dart
@@ -16,6 +16,7 @@ import '../src/datastore/fake_cocoon_config.dart';
 import '../src/request_handling/api_request_handler_tester.dart';
 import '../src/request_handling/fake_authentication.dart';
 import '../src/request_handling/fake_http.dart';
+import '../src/utils.dart';
 
 void main() {
   group('RefreshChromebotStatus', () {
@@ -134,6 +135,3 @@ void main() {
     });
   });
 }
-
-// ignore: must_be_immutable
-class MockLuciService extends Mock implements LuciService {}
