+++ app_dart/test/request_handlers/refresh_cirrus_status_test.dart
@@ -3,9 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:gcloud/db.dart';
-import 'package:github/github.dart';
 import 'package:graphql/client.dart';
-import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
 import 'package:cocoon_service/src/model/appengine/commit.dart';
@@ -213,7 +211,3 @@ QueryResult createQueryResult(List<dynamic> statuses, String branch) {
     },
   );
 }
-
-class MockGitHub extends Mock implements GitHub {}
-
-class MockRepositoriesService extends Mock implements RepositoriesService {}
